+++ spec/support/initialization_verifier.rb
@@ -1,0 +1,9 @@
+module InitializationVerifier
+  def initialize
+    @initialized = true
+  end
+
+  def initialized?
+    @initialized ||= false
+  end
+end